+++ app/Modules/Admin/Http/Controllers/Api/StatsController.php
@@ -12,23 +12,7 @@ class StatsController extends ApiController
     public function index()
     {
         try {
-
-            $totalUser = StatsServices::getUserCount();
-            $userGrobByRole = StatsServices::getUserCountByRole();
-            $totalJob = StatsServices::getJobCount();
-            $totlaApplicants = StatsServices::getApplicationCount();
-            $totalJobByType = StatsServices::getJobCountByType();
-            $countRecentRegistrations = StatsServices::getUserResentRegisteration();
-            return $this->respondWithSuccess([
-
-                "total_users" => $totalUser,
-                "total_jobs" =>  $totalJob,
-                "total_applications" => $totlaApplicants,
-                "users_by_role" => $userGrobByRole,
-                "jobs_by_type" => $totalJobByType,
-                "recent_registrations" => $countRecentRegistrations
-
-            ]);
+            return $this->respondWithSuccess(StatsServices::getAllStat());
         } catch (\Throwable $th) {
             return $this->respondError("ERROR  " .  $th->getMessage());
         }