+++ CHANGELOG.md
@@ -54,3 +54,4 @@
 
 - User staticties ( count by role , total , recent rejecter )
 - Job staticties ( count by type , total , total applicants )
+- AI Requests staticties ( count by status)